backend/requirements.txt
aniso8601==9.0.1
async-timeout==4.0.3
attrs==23.2.0
black==24.2.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
DBUtils==3.1.0
flasgger==0.9.7.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
greenlet==3.0.3
gunicorn==21.2.0
importlib_metadata==7.0.2
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mistune==3.0.2
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
referencing==0.34.0
rpds-py==0.18.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.10.0
waitress==3.0.0
Werkzeug==3.0.1
zipp==3.17.0
